refactor(backend): extract sales order lifecycle service
