+++ backend/src/modules/sales/sales.module.ts
@@ -32,6 +32,7 @@ import { InvoiceService } from './services/invoice.service';
 import { PaymentService } from './services/payment.service';
 import { SalesAnalyticsService } from './services/sales-analytics.service';
 import { InventoryIntegrationService } from './services/inventory-integration.service';
+import { SalesOrderLifecycleService } from './services/sales-order-lifecycle.service';
 import { SalesOrderPaymentService } from './services/sales-order-payment.service';
 import { TransactionManager } from '../../common/utils/transaction.util';
 
@@ -69,6 +70,7 @@ import { TransactionManager } from '../../common/utils/transaction.util';
     PaymentService,
     SalesAnalyticsService,
     InventoryIntegrationService,
+    SalesOrderLifecycleService,
     SalesOrderPaymentService,
     TransactionManager,
   ],
